+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## 0.2.2
 * Optimizer creation and update during training now part of DeepXubeNNet class. Methods can be overridden for different behavior.
+* step_min argument added to deepxube time
+* Added ReLU2 activation function
 
 ## 0.2.1
 * Consolidate search: Beam search -> special cases: greedy_policy, graph search -> special cases: batch weighted A* search, batch weighted Q* search